Added failing test for wradlib metadata to radolan header and refactored test files a little
