+++ radolan_to_netcdf/wradlib_to_radolan_bin.py
@@ -1,0 +1,20 @@
+import wradlib as wrl
+
+
+def metadata_to_header(metadata):
+    """
+
+    Parameters
+    ----------
+    metadata : dict
+        Dict of metadata as returned by wradlib.io.read_radolan_composite
+
+    Returns
+    -------
+
+    header : byte string
+        Header byte string conforming to the definition of RADOALN binary files
+
+    """
+
+    return None
